Create QUESTION 1
Booker University library maintains an inventory of books. The list includes the details: Author,  price, title, book_number and number of copies of each book. Whenever new books are purchased the librarian adds the book details into the  database. The chief librarian occasionally requests for a list of all the books in the database.
Required
Construct a simple class called book with suitable data members and member functions to:-
(i)	Insert a new book record into the database
(ii)	Display a list of all books in the database
(iii)	Write a main function to test the program