Create Question 5(b)
Employees of Fortune Business Systems Ltd contribute 5% of their monthly basic salary towards their individual pension savings. Since this may change in future, the operation to compute the pension has not been included. A friend function (to the class employee) called compute_pension() to calculate and return the pension contribution of an employee is therefore needed. In addition to the output as shown above.